Fixed no pid file exists error

+++ tika.js
@@ -79,31 +79,44 @@ var queue={
 var app={
 	"startServer":function(){
 		//first kill an old instance of tika if exists
-		try{
-			var pid=fs.readFileSync(__dirname+"/db/sqlite/tikaPID.txt").toString();
-			log.put("Trying to kill an old instance of tika if active","info");
-		}catch(err){
-			pid="";
-			log.put(err.stack,color_debug);
-			//touch file if not exists
-			var stream=fs.createWriteStream(__dirname+"/db/sqlite/tikaPID.txt");
-			stream.write("");
-			stream.end();
-		}
+		var pid = "";
+			try{
+					pid=fs.readFileSync(__dirname+"/db/sqlite/tikaPID.txt").toString();
+					log.put("Trying to kill an old instance of tika if active","info");
+			}catch(err){
+				//if file not exists
+					//touch file if not exists
+					var stream=fs.createWriteStream(__dirname+"/db/sqlite/tikaPID.txt");
+					stream.write("");
+					stream.end();
+			}
 		
 		
 		try{
-			process.kill(parseInt(pid));
+			if(pid !== ""){
+				process.kill(parseInt(pid));
+			}
+			
 		}
 		catch(err){
-
-			log.put(err.stack,color_debug);
+			if(err.code === "ESRCH"){
+				//cannot reach the process by the pid
+				//maybe process is already killed
+				log.put("Old tika instance was already killed","info");
+				log.put("Reset tika pid file","info");
+				var stream=fs.createWriteStream(__dirname+"/db/sqlite/tikaPID.txt");
+				stream.write("");
+				stream.end();
+			}else{
+				log.put(err.stack,color_debug,err.type);
+			}
+			
 		}
 		var d=exec('java -jar '+__dirname+'/lib/tika-server-1.11.jar -h '+config.getConfig("tika_host"), function(error, stdout, stderr) {
 			log.put("[SUCCESS] Tika server started","success");
 		    if (check.assigned(error)) {
 		    	log.put(error.stack,color_debug);
-		        log.put('[INFO] Server is already running',"info");
+		        log.put('Server is already running',"info");
 		    }
 		});
 		try{
